+++ src/net/wiiu/VideoHandlerWiiU.h
@@ -22,7 +22,7 @@ private:
     uint32_t frame_index;
     uint8_t image_buf[2000000];
 
-    int h264_nal_encapsulate(bool is_idr, uint8_t *frame, size_t frame_size, uint8_t *nals);
+    size_t h264_nal_encapsulate(bool is_idr, uint8_t *frame, size_t frame_size, uint8_t *nals);
 
     void update(unsigned char *packet, size_t packet_size, sockaddr_in *from_address,
                 unsigned int *address_size) override;
